SONARHTML-384 Move ERB extension collection out of HtmlSensor
Building the recognized-extensions set and composing the ERB file
predicate are concerns of the filter, not the sensor. HtmlSensor was
accumulating helpers it did not need to know about.

- HtmlConstants.OTHER_FILE_SUFFIXES now lives alongside the other
  extension constants instead of as a private array on HtmlSensor.
- ErbFileFilter exposes filePredicate(FilePredicates, Configuration)
  and owns recognizedExtensions(Configuration) + addExtensions.
- HtmlSensor.execute calls ErbFileFilter.filePredicate(...) and no
  longer carries collectRecognizedExtensions, addExtensions, or
  erbContentPredicate.
